Implemented deletion of CSR. Added confirmation dialogues.

+++ admin-backend/src/main/java/com/example/demo/controller/CertificatesController.java
@@ -86,7 +86,7 @@ public class CertificatesController {
     }
 
     @DeleteMapping(value = "/csr/{id}")
-    public void deleteCSR(@PathVariable(name="id") Long id){
+    public void deleteCSR(@PathVariable(name="id") Integer id){
         this.certificatesService.deleteCsr(id);
     }
